`@remotion/media-parser`: Name functions better
